refactor: remove constants in upper scope
