+++ src/core/constants/amountOfWidthPoints.ts
